requirements.txt
langserve[all]
langchain
langchain_groq
fastapi
uvicorn
python-dotenv
pydantic